+++ manta-pay/src/key.rs
@@ -29,13 +29,14 @@ use core::marker::PhantomData;
 use manta_accounting::key::{
     self, AccountIndex, HierarchicalKeyDerivationScheme, IndexType, KeyIndex, Kind,
 };
-use manta_crypto::rand::{CryptoRng, RngCore, Sample};
+use manta_crypto::rand::{CryptoRng, RngCore};
 use manta_util::{create_seal, seal, Array};
 
 #[cfg(feature = "serde")]
 use manta_util::serde::{Deserialize, Serialize, Serializer};
 
-pub use bip32::{Error, Seed, XPrv as SecretKey};
+pub use bip0039;
+pub use bip32::{self, Error, XPrv as SecretKey};
 
 create_seal! {}
 
@@ -127,14 +128,14 @@ impl_coin_type!(
 /// Account Table Type
 pub type AccountTable<C> = key::AccountTable<KeySecret<C>>;
 
-/// Seed Byte Array Type
-type SeedBytes = Array<u8, { Seed::SIZE }>;
+/// Seed Bytes
+pub type SeedBytes = Array<u8, { bip32::Seed::SIZE }>;
 
 /// Key Secret
 #[cfg_attr(
     feature = "serde",
     derive(Deserialize, Serialize),
-    serde(crate = "manta_util::serde", deny_unknown_fields, transparent)
+    serde(crate = "manta_util::serde", deny_unknown_fields)
 )]
 #[derive(derivative::Derivative)]
 #[derivative(Clone(bound = ""))]
@@ -145,6 +146,9 @@ where
     /// Key Seed
     seed: SeedBytes,
 
+    /// Mnemonic
+    mnemonic: Mnemonic,
+
     /// Type Parameter Marker
     __: PhantomData<C>,
 }
@@ -153,41 +157,36 @@ impl<C> KeySecret<C>
 where
     C: CoinType,
 {
-    /// Builds a [`KeySecret`] from raw bytes.
+    /// Builds a [`KeySecret`] from `seed` and `mnemonic`.
     #[inline]
-    fn build(seed: [u8; Seed::SIZE]) -> Self {
+    fn new_unchecked(seed: [u8; bip32::Seed::SIZE], mnemonic: Mnemonic) -> Self {
         Self {
             seed: seed.into(),
+            mnemonic,
             __: PhantomData,
         }
-    }
-
-    /// Builds a [`KeySecret`] from a `seed`.
-    #[inline]
-    fn from_seed(seed: Seed) -> Self {
-        Self::build(*seed.as_bytes())
     }
 
     /// Converts a `mnemonic` phrase into a [`KeySecret`], locking it with `password`.
     #[inline]
     #[must_use]
     pub fn new(mnemonic: Mnemonic, password: &str) -> Self {
-        Self::from_seed(mnemonic.to_seed(password))
+        Self::new_unchecked(mnemonic.to_seed(password), mnemonic)
     }
-}
 
-impl<C> Sample for KeySecret<C>
-where
-    C: CoinType,
-{
+    /// Exposes a shared reference to the [`Mnemonic`] for `self`.
     #[inline]
-    fn sample<R>(_: (), rng: &mut R) -> Self
+    pub fn expose_mnemonic(&self) -> &Mnemonic {
+        &self.mnemonic
+    }
+
+    /// Samples a random [`KeySecret`] from `rng` with no password.
+    #[inline]
+    pub fn sample<R>(rng: &mut R) -> Self
     where
-        R: RngCore + ?Sized,
+        R: CryptoRng + RngCore + ?Sized,
     {
-        let mut seed = [0; Seed::SIZE];
-        rng.fill_bytes(&mut seed);
-        Self::build(seed)
+        Self::new(Mnemonic::sample(rng), "")
     }
 }
 
@@ -241,26 +240,28 @@ where
 pub struct Mnemonic(
     /// Underlying BIP39 Mnemonic
     #[cfg_attr(feature = "serde", serde(serialize_with = "Mnemonic::serialize"))]
-    bip32::Mnemonic,
+    bip0039::Mnemonic,
 );
 
+/// Seed Type
+pub type Seed = [u8; 64];
+
 impl Mnemonic {
-    /// Create a new BIP39 mnemonic phrase from the given string.
+    /// Create a new BIP39 mnemonic phrase from the given phrase.
     #[inline]
-    pub fn new<S>(phrase: S) -> Result<Self, Error>
-    where
-        S: AsRef<str>,
-    {
-        bip32::Mnemonic::new(phrase, Default::default()).map(Self)
+    pub fn new(phrase: &str) -> Result<Self, Error> {
+        Ok(Self(bip0039::Mnemonic::from_phrase(phrase).unwrap()))
     }
 
-    /// Samples a random [`Mnemonic`] using the entropy returned from `rng`.
+    /// Samples a random 12 word [`Mnemonic`] using the entropy returned from `rng`.
     #[inline]
     pub fn sample<R>(rng: &mut R) -> Self
     where
         R: CryptoRng + RngCore + ?Sized,
     {
-        Self(bip32::Mnemonic::random(rng, Default::default()))
+        let mut entropy: [u8; 16] = [0; 16];
+        rng.fill_bytes(&mut entropy);
+        Self(bip0039::Mnemonic::from_entropy(entropy.to_vec()).unwrap())
     }
 
     /// Convert this mnemonic phrase into the BIP39 seed value.
@@ -272,7 +273,7 @@ impl Mnemonic {
     /// Serializes the underlying `mnemonic` phrase.
     #[cfg(feature = "serde")]
     #[inline]
-    fn serialize<S>(mnemonic: &bip32::Mnemonic, serializer: S) -> Result<S::Ok, S::Error>
+    fn serialize<S>(mnemonic: &bip0039::Mnemonic, serializer: S) -> Result<S::Ok, S::Error>
     where
         S: Serializer,
     {
@@ -301,6 +302,6 @@ impl TryFrom<String> for Mnemonic {
 
     #[inline]
     fn try_from(string: String) -> Result<Self, Self::Error> {
-        Self::new(string)
+        Self::new(string.as_str())
     }
 }