+++ manta-pay/src/signer/client/mod.rs
@@ -14,12 +14,12 @@
 // You should have received a copy of the GNU General Public License
 // along with manta-rs.  If not, see <http://www.gnu.org/licenses/>.
 
-//! Signer Client Implementations
-
-#[cfg(feature = "http")]
-#[cfg_attr(doc_cfg, doc(cfg(feature = "http")))]
-pub mod http;
+//! Signer Client
 
 #[cfg(feature = "websocket")]
 #[cfg_attr(doc_cfg, doc(cfg(feature = "websocket")))]
 pub mod websocket;
+
+#[cfg(feature = "network")]
+#[cfg_attr(doc_cfg, doc(cfg(feature = "network")))]
+pub mod network;